refactor(starknet_http_server): move metrics test to own module
commit-id:c94d8d6e
